refactor: move shikimori services to specific dir
